create text-based common datasource
